\name{CD_longitudinal}
\alias{CD_longitudinal}
\docType{data}
\title{
A longitudinal dataset of real microbiome
}
\description{
This is a longitudinal dataset of real microbiome.
}
\usage{data(CD_longitudinal)}
\format{
phyloseq
}

\references{
McMurdie PJ and Holmes S. phyloseq: An R package for reproducible interactive analysis and graphics of microbiome census data. PLoS ONE. 2013;8(4):e61217

Vázquez-Baeza Y, et al. Guiding longitudinal sampling in IBD cohorts. _Gut_ 2018;67(9):1743-1745
}

\examples{
library(phyloseq)

data(CD_longitudinal)
CD_longitudinal


}
